% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brute_force.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute Force}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{A data frame with numeric positive values.}

\item{W}{A positive numeric scalar.}
}
\value{
A list with optimal value, weight and selected objects.
}
\description{
Gives you the optimal solution with the Knapsack problem computed by brute force.
Further information about the knapsack problem \url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
