% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_heuristic.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Heuristic}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A data frame with numeric positive values.}

\item{W}{A positive numeric scalar.}
}
\value{
A list approximate of optimal value, weight and selected objects.
}
\description{
Gives you an approxiamtion of the optimal solution with the Knapsack problem computed by greedy heuristic
Further information about the knapsack problem \url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
