\name{knapsack}

\alias{knapsack-package}
\docType{package}
\title{
\packageTitle{knapsack}
}
\description{
\packageDescription{knapsack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{knapsack}
\packageIndices{knapsack}

}
\author{
\packageAuthor{knapsack}

Maintainer: \packageMaintainer{knapsack}
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}

\keyword{ package }
\seealso{
\url{https://github.com/poceviciute/AdvRprogr_lab6}

}
\examples{

}
