% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Knapsack Dynamic Algorithm}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{A data frame with numeric positive values.}

\item{W}{A positive numeric scalar.}
}
\value{
A list with maximum value and elements.
}
\description{
The knapsack problem is a problem in combinatorial optimization: how to maximasiethe value of fitted object by fullfilling the weight constraint.
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
