\name{knapsack_objects}
\alias{knapsack_objects}
\docType{data}
\title{
knapsack_object
}
\description{
knapsack_object
}
\usage{data("knapsack_objects")}
\format{
  A data frame with 2000 observations on the following 2 variables.
  \describe{
    \item{\code{w}}{a numeric vector}
    \item{\code{v}}{a numeric vector}
  }
}
\details{
A data frame with 2000 observations on the following 2 variables.
}

\examples{
data(knapsack_objects)
## maybe str(knapsack_objects) ; plot(knapsack_objects) ...
}
\keyword{datasets}
